+++ event/bus/locals.tf
@@ -1,12 +1,16 @@
 module "name_map" {
-  source   = "../../name_map"
-  name_map = var.bus_map
-  std_map  = var.std_map
+  source                          = "../../name_map"
+  name_append_default             = var.name_append_default
+  name_include_app_fields_default = var.name_include_app_fields_default
+  name_infix_default              = var.name_infix_default
+  name_map                        = local.l0_map
+  name_prepend_default            = var.name_prepend_default
+  std_map                         = var.std_map
 }
 
 module "policy_map" {
   source                      = "../../iam/policy/name_map"
-  name_map                    = var.bus_map
+  name_map                    = local.l0_map
   policy_access_list_default  = var.policy_access_list_default
   policy_create_default       = var.policy_create_default
   policy_name_append_default  = var.policy_name_append_default
@@ -22,35 +26,39 @@ locals {
     for k, v in local.event_bus_map : k => v if v.archive_retention_days != null
   }
   create_bus_map = {
-    for k, v in local.lx_map : k => v if v.event_bus_name == null
+    for k, v in local.lx_map : k => v if v.create_bus
   }
   create_bus_data_map = {
-    for k, v in local.lx_map : k => v if v.event_bus_name != null
+    for k, v in local.lx_map : k => v if !v.create_bus
   }
   create_policy_map = {
     for k, v in local.event_bus_map : k => v if v.policy_create
   }
   event_bus_map = {
     for k, v in local.lx_map : k => merge(v, {
-      event_bus_arn  = v.event_bus_name == null ? aws_cloudwatch_event_bus.this_bus[k].arn : data.aws_cloudwatch_event_bus.this_bus[k].arn
-      event_bus_name = v.event_bus_name == null ? aws_cloudwatch_event_bus.this_bus[k].name : v.event_bus_name
+      event_bus_arn  = v.create_bus ? aws_cloudwatch_event_bus.this_bus[k].arn : data.aws_cloudwatch_event_bus.this_bus[k].arn
+      event_bus_name = v.create_bus ? aws_cloudwatch_event_bus.this_bus[k].name : v.event_bus_name
     })
   }
+  l0_map = {
+    for k, v in var.bus_map : k => v
+  }
   l1_map = {
-    for k, v in var.bus_map : k => merge(v, module.name_map.data[k], module.policy_map.data[k], {
+    for k, v in local.l0_map : k => merge(v, module.name_map.data[k], module.policy_map.data[k], {
       archive_retention_days        = v.archive_retention_days == null ? var.bus_archive_retention_days_default : v.archive_retention_days
+      create_bus                    = v.event_bus_name == null
       logging_enabled               = v.logging_enabled == null ? var.bus_logging_enabled_default : v.logging_enabled
       logging_excluded_detail_types = v.logging_excluded_detail_types == null ? var.bus_logging_excluded_detail_types_default : v.logging_excluded_detail_types
       log_retention_days            = v.log_retention_days == null ? var.bus_log_retention_days_default : v.log_retention_days
     })
   }
   l2_map = {
-    for k, v in var.bus_map : k => {
+    for k, v in local.l0_map : k => {
       log_name = "bus_${k}"
     }
   }
   lx_map = {
-    for k, v in var.bus_map : k => merge(local.l1_map[k], local.l2_map[k])
+    for k, v in local.l0_map : k => merge(local.l1_map[k], local.l2_map[k])
   }
   log_map = {
     for k, v in local.lx_map : v.log_name => merge(v, {
@@ -69,6 +77,9 @@ locals {
       bus = merge(
         v,
         module.bus_policy[k].data,
+        {
+          iam_policy_doc = v.policy_create ? module.resource_policy[k].iam_policy_doc : null
+        },
       )
       log     = module.log_group.data[v.log_name]
       trigger = module.log_trigger.data[v.log_name]